refactor: Updated Services.

+++ src/main/java/com/projects/Student_Information_System/Service/StudentService.java
@@ -1,11 +1,14 @@
 package com.projects.Student_Information_System.Service;
 
+import com.projects.Student_Information_System.Model.DTO.CourseDTO;
 import com.projects.Student_Information_System.Model.DTO.StudentDTO;
 import com.projects.Student_Information_System.Model.Enums.Role;
 import com.projects.Student_Information_System.Model.Student;
 import com.projects.Student_Information_System.Repository.IStudentRepository;
 import com.projects.Student_Information_System.Util.IDGenerator;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -23,17 +26,30 @@ public class StudentService {
         this.studentRepository = studentRepository;
     }
 
-    public List<StudentDTO> getAllStudents() {
-        return studentRepository.findAll()
+    public List<StudentDTO> getAllStudents(int page, int size) {
+        Pageable pageable = PageRequest.of(page, size);
+        return studentRepository.findAll(pageable)
                 .stream()
                 .map(this::convertStudentEntityToDTO)
-                .collect(Collectors.toList());
+                .toList();
     }
 
     public StudentDTO getStudent(Long id) {
         return convertStudentEntityToDTO(findStudentById(id));
     }
 
+    public List<CourseDTO> getStudentSubjects(Long id) {
+        return findStudentById(id).getEnrolledCourses()
+                .stream()
+                .map(course -> {
+                    return new CourseDTO().builder()
+                            .name(course.getName())
+                            .code(course.getCode())
+                            .description(course.getDescription())
+                            .build();
+                })
+                .toList();
+    }
     @Transactional
     public StudentDTO createStudent(StudentDTO studentDTO) {
         var student = convertStudentDTOToEntity(studentDTO);
@@ -45,13 +61,16 @@ public class StudentService {
     @Transactional
     public StudentDTO updateStudent(Long id, StudentDTO studentDTO) {
         var studentToUpdate = findStudentById(id);
-        studentToUpdate.setPersonalInformation(studentDTO.getPersonalInformation());
+        studentToUpdate.setPersonalInformation((studentDTO.getPersonalInformation() != null) ? studentDTO.getPersonalInformation() : studentToUpdate.getPersonalInformation());
         return convertStudentEntityToDTO(studentRepository.save(studentToUpdate));
     }
 
     @Transactional
     public void deleteStudent(Long id) {
         var studentToDelete = findStudentById(id);
+        if (!studentToDelete.getEnrolledCourses().isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Student is enrolled in a course");
+        }
         studentRepository.deleteById(studentToDelete.getId());
     }
 
@@ -63,9 +82,10 @@ public class StudentService {
     }
 
     private Long getLatestCount() {
-        return getAllStudents()
+        List<Student> students = studentRepository.findAll();
+        return students
                 .stream()
-                .mapToLong(StudentDTO::getId).max().orElse(0L);
+                .mapToLong(Student::getId).max().orElse(0L);
     }
 
     public StudentDTO convertStudentEntityToDTO(Student student) {
@@ -73,6 +93,7 @@ public class StudentService {
         studentDTO.setId(student.getId());
         studentDTO.setStudentId(student.getStudentId());
         studentDTO.setRole(student.getRole());
+        studentDTO.setUsername(student.getUsername());
         studentDTO.setPersonalInformation(student.getPersonalInformation());
         return studentDTO;
     }
